refactor: full refactor of the function
